+++ core/java/androidx/test/core/view/ViewCapture.kt
@@ -29,22 +29,27 @@ import android.view.PixelCopy
 import android.view.Surface
 import android.view.SurfaceView
 import android.view.View
-import android.view.ViewTreeObserver
+import android.view.ViewTreeObserver.OnDrawListener
 import android.view.WindowManager
 import androidx.annotation.RequiresApi
-import androidx.concurrent.futures.ResolvableFuture
+import androidx.concurrent.futures.SuspendToFutureAdapter
 import androidx.test.annotation.ExperimentalTestApi
 import androidx.test.core.internal.os.HandlerExecutor
 import androidx.test.internal.platform.ServiceLoaderWrapper
 import androidx.test.internal.platform.os.ControlledLooper
 import androidx.test.internal.platform.reflect.ReflectiveField
 import androidx.test.internal.platform.reflect.ReflectiveMethod
+import androidx.test.internal.util.Checks.checkState
 import androidx.test.platform.graphics.HardwareRendererCompat
 import com.google.common.util.concurrent.ListenableFuture
 import java.util.function.Consumer
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.suspendCancellableCoroutine
 
 /**
- * Asynchronously captures an image of the underlying view into a [Bitmap].
+ * Suspend function for capturing an image of the underlying view into a [Bitmap].
  *
  * For devices below [Build.VERSION_CODES#O] (or if the view's window cannot be determined), the
  * image is obtained using [View#draw]. Otherwise, [PixelCopy] is used.
@@ -52,39 +57,44 @@ import java.util.function.Consumer
  * This method will also enable [HardwareRendererCompat#setDrawingEnabled(boolean)] if required.
  *
  * This API is primarily intended for use in lower layer libraries or frameworks. For test authors,
- * its recommended to use espresso or compose's captureToImage.
+ * it's recommended to use Espresso or Compose's captureToImage.
  *
  * If a rect is supplied, this will further crop locally from the bounds of the given view. For
  * example, if the given view is at (10, 10 - 30, 30) and the rect is (5, 5 - 10, 10), the final
  * bitmap will be a 5x5 bitmap that spans (15, 15 - 20, 20). This is particularly useful for
  * Compose, which only has a singular view that contains a hierarchy of nodes.
  *
- * Note: This API will likely change in the future to a kotlin suspend function. If you want a
- * method that returns a ListenableFuture instead, use [captureToBitmapAsync].
+ * This API must be called on the View's handler thread. If you're calling this from another
+ * context, eg directly from the test thread, you can use something like
+ * <pre>{@code
+ *   runBlocking(view.handler.asCoroutineDispatcher()) {
+ *     withTimeout(10.seconds) {
+ *       view.captureToBitmap(rect)
+ *     }
+ *   }
+ * }</pre>
  *
  * This API is currently experimental and subject to change or removal.
  */
 @ExperimentalTestApi
-fun View.captureToBitmap(rect: Rect? = null): ListenableFuture<Bitmap> {
-  val bitmapFuture: ResolvableFuture<Bitmap> = ResolvableFuture.create()
-  val mainExecutor = HandlerExecutor(Handler(Looper.getMainLooper()))
+suspend fun View.captureToBitmap(rect: Rect? = null): Bitmap {
+  checkState(isAttachedToWindow, "View must be attached to a window")
+  checkState(
+    handler.looper.isCurrentThread,
+    "Must be called from view's handler thread. Current: ${Thread.currentThread().name}, view handler: ${handler.looper.thread.name}",
+  )
 
-  // disable drawing again if necessary once work is complete
-  if (!HardwareRendererCompat.isDrawingEnabled()) {
-    HardwareRendererCompat.setDrawingEnabled(true)
-    bitmapFuture.addListener({ HardwareRendererCompat.setDrawingEnabled(false) }, mainExecutor)
+  var bitmap: Bitmap? = null
+  val hardwareDrawingEnabled = HardwareRendererCompat.isDrawingEnabled()
+  HardwareRendererCompat.setDrawingEnabled(true)
+  try {
+    forceRedraw()
+    bitmap = generateBitmap(rect)
+  } finally {
+    HardwareRendererCompat.setDrawingEnabled(hardwareDrawingEnabled)
   }
 
-  mainExecutor.execute {
-    if (getControlledLooper().areDrawCallbacksSupported()) {
-      val forceRedrawFuture = forceRedraw()
-      forceRedrawFuture.addListener({ generateBitmap(bitmapFuture, rect) }, mainExecutor)
-    } else {
-      generateBitmap(bitmapFuture, rect)
-    }
-  }
-
-  return bitmapFuture
+  return bitmap!!
 }
 
 private fun getControlledLooper(): ControlledLooper {
@@ -93,87 +103,90 @@ private fun getControlledLooper(): ControlledLooper {
   }
 }
 
-/**
- * Equivalent to [captureToBitmap] for now. In the future [captureToBitmap] will be changed to a
- * suspend function.
- */
+/** A ListenableFuture variant of captureToBitmap intended for use from Java. */
 @ExperimentalTestApi
 fun View.captureToBitmapAsync(rect: Rect? = null): ListenableFuture<Bitmap> {
-  return captureToBitmap(rect)
+  return SuspendToFutureAdapter.launchFuture(Dispatchers.Main) { captureToBitmap(rect) }
 }
 
 /**
  * Trigger a redraw of the given view.
  *
  * Should only be called on UI thread.
- *
- * @return a [ListenableFuture] that will be complete once ui drawing is complete
  */
 // TODO(b/316921934): uncomment once @ExperimentalTestApi is removed
 // @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @ExperimentalTestApi
-fun View.forceRedraw(): ListenableFuture<Void> {
-  val future: ResolvableFuture<Void> = ResolvableFuture.create()
-
-  if (Build.VERSION.SDK_INT >= 29 && isHardwareAccelerated) {
-    viewTreeObserver.registerFrameCommitCallback() { future.set(null) }
-  } else {
-    viewTreeObserver.addOnDrawListener(
-      object : ViewTreeObserver.OnDrawListener {
-        var handled = false
-
-        override fun onDraw() {
-          if (!handled) {
-            handled = true
-            future.set(null)
-            // cannot remove on draw listener inside of onDraw
-            Handler(Looper.getMainLooper()).post { viewTreeObserver.removeOnDrawListener(this) }
-          }
-        }
-      }
-    )
-  }
-  invalidate()
-  return future
-}
-
-private fun View.generateBitmap(bitmapFuture: ResolvableFuture<Bitmap>, rect: Rect? = null) {
-  if (bitmapFuture.isCancelled) {
+suspend fun View.forceRedraw() {
+  checkState(handler.looper.isCurrentThread, "Must be called from view's handler thread")
+  if (!getControlledLooper().areDrawCallbacksSupported()) {
+    Log.i("ViewCapture", "Skipping forceRedraw as it is not supported")
     return
   }
+
+  var drawListener: OnDrawListener? = null
+  try {
+    return suspendCancellableCoroutine<Unit> { cont ->
+      if (Build.VERSION.SDK_INT >= 29 && isHardwareAccelerated) {
+        viewTreeObserver.registerFrameCommitCallback() { cont.resume(Unit) }
+      } else {
+        drawListener =
+          object : OnDrawListener {
+            var handled = false
+
+            override fun onDraw() {
+              if (!handled) {
+                handled = true
+
+                cont.resume(Unit)
+              }
+            }
+          }
+        viewTreeObserver.addOnDrawListener(drawListener)
+      }
+      invalidate()
+    }
+  } finally {
+    // cannot remove on draw listener inside of onDraw
+    if (drawListener != null) {
+
+      viewTreeObserver.removeOnDrawListener(drawListener)
+    }
+  }
+}
+
+private suspend fun View.generateBitmap(rect: Rect? = null): Bitmap {
   val rectWidth = rect?.width() ?: width
   val rectHeight = rect?.height() ?: height
   val destBitmap = Bitmap.createBitmap(rectWidth, rectHeight, Bitmap.Config.ARGB_8888)
-  when {
-    Build.VERSION.SDK_INT < 26 -> generateBitmapFromDraw(destBitmap, bitmapFuture, rect)
-    Build.VERSION.SDK_INT >= 34 -> generateBitmapFromPixelCopy(destBitmap, bitmapFuture, rect)
-    this is SurfaceView -> generateBitmapFromSurfaceViewPixelCopy(destBitmap, bitmapFuture, rect)
-    else -> generateBitmapFromPixelCopy(this.getSurface(), destBitmap, bitmapFuture, rect)
+
+  return when {
+    Build.VERSION.SDK_INT < 26 -> generateBitmapFromDraw(destBitmap, rect)
+    Build.VERSION.SDK_INT >= 34 -> generateBitmapFromPixelCopy(destBitmap, rect)
+    this is SurfaceView -> generateBitmapFromSurfaceViewPixelCopy(destBitmap, rect)
+    else -> generateBitmapFromPixelCopy(this.getSurface(), destBitmap, rect)
   }
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
-private fun SurfaceView.generateBitmapFromSurfaceViewPixelCopy(
+private suspend fun SurfaceView.generateBitmapFromSurfaceViewPixelCopy(
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
   rect: Rect?,
-) {
-  val onCopyFinished =
-    PixelCopy.OnPixelCopyFinishedListener { result ->
-      if (result == PixelCopy.SUCCESS) {
-        bitmapFuture.set(destBitmap)
-      } else {
-        bitmapFuture.setException(RuntimeException(String.format("PixelCopy failed: %d", result)))
+): Bitmap {
+  return suspendCancellableCoroutine<Bitmap> { cont ->
+    val onCopyFinished =
+      PixelCopy.OnPixelCopyFinishedListener { result ->
+        if (result == PixelCopy.SUCCESS) {
+          cont.resume(destBitmap)
+        } else {
+          cont.resumeWithException(RuntimeException(String.format("PixelCopy failed: %d", result)))
+        }
       }
-    }
-  PixelCopy.request(this, rect, destBitmap, onCopyFinished, handler)
+    PixelCopy.request(this, rect, destBitmap, onCopyFinished, handler)
+  }
 }
 
-internal fun View.generateBitmapFromDraw(
-  destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
-  rect: Rect?,
-) {
+internal fun View.generateBitmapFromDraw(destBitmap: Bitmap, rect: Rect?): Bitmap {
   destBitmap.density = resources.displayMetrics.densityDpi
   computeScroll()
   val canvas = Canvas(destBitmap)
@@ -183,7 +196,7 @@ internal fun View.generateBitmapFromDraw(
   }
 
   draw(canvas)
-  bitmapFuture.set(destBitmap)
+  return destBitmap
 }
 
 /**
@@ -193,32 +206,32 @@ internal fun View.generateBitmapFromDraw(
  * [PixelCopy.ofWindow(View)], and will be called when running on Android API levels O to T.
  */
 @RequiresApi(Build.VERSION_CODES.O)
-private fun View.generateBitmapFromPixelCopy(
+private suspend fun View.generateBitmapFromPixelCopy(
   surface: Surface,
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
   rect: Rect?,
-) {
-  val onCopyFinished =
-    PixelCopy.OnPixelCopyFinishedListener { result ->
-      if (result == PixelCopy.SUCCESS) {
-        bitmapFuture.set(destBitmap)
-      } else {
-        bitmapFuture.setException(RuntimeException("PixelCopy failed: $result"))
-      }
+): Bitmap {
+  return suspendCancellableCoroutine<Bitmap> { cont ->
+    var bounds = getBoundsInSurface()
+    if (rect != null) {
+      bounds =
+        Rect(
+          bounds.left + rect.left,
+          bounds.top + rect.top,
+          bounds.left + rect.right,
+          bounds.top + rect.bottom,
+        )
     }
-
-  var bounds = getBoundsInSurface()
-  if (rect != null) {
-    bounds =
-      Rect(
-        bounds.left + rect.left,
-        bounds.top + rect.top,
-        bounds.left + rect.right,
-        bounds.top + rect.bottom,
-      )
+    val onCopyFinished =
+      PixelCopy.OnPixelCopyFinishedListener { result ->
+        if (result == PixelCopy.SUCCESS) {
+          cont.resume(destBitmap)
+        } else {
+          cont.resumeWithException(RuntimeException("PixelCopy failed: $result"))
+        }
+      }
+    PixelCopy.request(surface, bounds, destBitmap, onCopyFinished, Handler(Looper.getMainLooper()))
   }
-  PixelCopy.request(surface, bounds, destBitmap, onCopyFinished, Handler(Looper.getMainLooper()))
 }
 
 /** Returns the Rect indicating the View's coordinates within its containing window. */
@@ -281,27 +294,24 @@ private fun View.reflectivelyGetLocationInSurface(locationInSurface: IntArray) {
   }
 }
 
-/** Generates a bitmap given the current view using android U's [PixelCopy.ofWindow(View)]. */
-@RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
-private fun View.generateBitmapFromPixelCopy(
+private suspend fun View.generateBitmapFromPixelCopy(
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
-  rect: Rect?,
-) {
-  val request =
-    PixelCopy.Request.Builder.ofWindow(this)
-      .setSourceRect(rect ?: getBoundsInWindow())
-      .setDestinationBitmap(destBitmap)
-      .build()
-  val mainExecutor = HandlerExecutor(Handler(Looper.getMainLooper()))
-
-  val onCopyFinished =
-    Consumer<PixelCopy.Result> { result ->
-      if (result.status == PixelCopy.SUCCESS) {
-        bitmapFuture.set(result.bitmap)
-      } else {
-        bitmapFuture.setException(RuntimeException("PixelCopy failed: $(result.status)"))
+  rect: Rect? = null,
+): Bitmap {
+  return suspendCancellableCoroutine<Bitmap> { cont ->
+    val request =
+      PixelCopy.Request.Builder.ofWindow(this)
+        .setSourceRect(rect ?: getBoundsInWindow())
+        .setDestinationBitmap(destBitmap)
+        .build()
+    val onCopyFinished =
+      Consumer<PixelCopy.Result> { result ->
+        if (result.status == PixelCopy.SUCCESS) {
+          cont.resume(result.bitmap)
+        } else {
+          cont.resumeWithException(RuntimeException("PixelCopy failed: $(result.status)"))
+        }
       }
-    }
-  PixelCopy.request(request, mainExecutor, onCopyFinished)
+    PixelCopy.request(request, HandlerExecutor(handler), onCopyFinished)
+  }
 }