+++ core/CHANGELOG.md
@@ -15,8 +15,10 @@
 **API Changes**
 
 * Added ApplicationInfoBuilder.setFlags(int)
+* Make suspend function versions of ViewCapture/WindowCapture/DeviceCapture APIs
 * Make Bitmap.writeToTestStorage use the registered PlatformTestStorage instead of hardcoding TestStorage
 * Add *Async variants of capture*ToBitmap methods
+
 
 **Breaking API Changes**